Update code based on review
Update intention of checkboxUpdate -> refocusEditorAndCursor
Move scroll editor into line method to codemirror util
Refactor initial cursor method call for refocus
