+++ src/components/atoms/CustomizedMarkdownPreviewer.tsx
@@ -9,7 +9,7 @@ interface CustomizedMarkdownPreviewer {
   attachmentMap?: ObjectMap<Attachment>
   updateContent?: (
     newContentOrUpdater: string | ((newValue: string) => string),
-    checkboxUpdate?: boolean
+    refocusEditorAndCursor?: boolean
   ) => void
 }
